+++ CHANGELOG.md
@@ -2,8 +2,12 @@
 
 ## Unreleased
 
-* `signing.sign_get_url()` is deprecated and will be removed in a future release.
-* `laterpay.signing.sign_and_encode()` is deprecated. Consider using `laterpay.utils.signed_query()` instead.
+* `laterpay.signing.sign_and_encode()` is deprecated and will be removed in a
+  future release. Consider using `laterpay.utils.signed_query()` instead.
+* `signing.sign_get_url()` is deprecated and will be removed in a future
+  release.
+* `LaterPayClient.get_access()` is deprecated and will be removed in a future
+  release. Consider using `LaterPayClient.get_access_data()` instead.
 
 
 ## 4.1.0